+++ src/app/g/[groupId]/page.tsx
@@ -13,9 +13,7 @@ const Page = async ({ params }: { params: Promise<{ groupId: string }> }) => {
   const groupData = await fetchGroupData(groupId)
   if (!groupData) return null
 
-  // todo: update this on RELEASE
   const roadmap = await fetchRoadmap(groupData.group_progress)
-  // const roadmap = await fetchRoadmap([])
 
   // if the group does not start yet, return message
   if (roadmap.length === 0) {